requirements.txt
# Core Dependencies
streamlit==1.31.0
langchain==0.1.10
langchain-openai==0.0.8
langchain-community==0.0.25

# Vector Store & Embeddings
chromadb==0.4.22
faiss-cpu==1.7.4

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0

# Knowledge Graph
networkx==3.2.1

# Text Processing
tiktoken==0.6.0
openai==1.12.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
numpy==1.26.4